+++ change_log.md
@@ -1,3 +1,4 @@
-- time input displayed correctly
-- departure and destination made lowercase when sent to backend
-- time sent in correct format to backend
+- departure and destination first letter capitalized, NOTE: not in previous commit but this one
+- limiting ordering fields with ride resource, NOTE: not in previous commit but this one
+- adding time input to tests, tests pass
+